型安全性の向上と参照返しの最適化を実施
各所でstatic_castによる型変換を追加し、float型での計算を明示化しました。WorldTransformの各ゲッターを参照返しに変更し、不要なコピーを削減。Enemyクラスにダミーキャラクターの回転・位置取得メソッドを追加。ModelEdgeEmitterでlineCount_の型変換も適用し、型安全性を強化しました。

+++ application/camera/title/AlignCameraState.cpp
@@ -15,8 +15,8 @@ void AlignCameraState::Update(float deltaTime)
 	float diff = targetRotate.y - cameraRotate.y;
 
 	// -π ～ π に正規化（最短差分にする）
-	while (diff > std::numbers::pi)  diff -= 2.0f * std::numbers::pi;
-	while (diff < -std::numbers::pi) diff += 2.0f * std::numbers::pi;
+	while (diff > static_cast<float>(std::numbers::pi))  diff -= 2.0f * static_cast<float>(std::numbers::pi);
+	while (diff < -static_cast<float>(std::numbers::pi)) diff += 2.0f * static_cast<float>(std::numbers::pi);
 
 	// 最短方向にイージング
 	cameraRotate.y += diff * 0.04f;
