+++ application/camera/title/RotateCameraState.cpp
@@ -13,9 +13,9 @@ void RotateCameraState::Update(float deltaTime)
 
 	rotate.y += kRotateSpeed_ * deltaTime;
 
-	if(rotate.y > std::numbers::pi * 2)
+	if(rotate.y > static_cast<float>(std::numbers::pi * 2))
 	{
-		rotate.y -= std::numbers::pi * 2;
+		rotate.y -= static_cast<float>(std::numbers::pi * 2);
 	}
 
 	camera_->GetWorldTransform().SetRotate(rotate);
